requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==2.0.0
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
cssselect2==0.8.0
daphne==4.2.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==4.2.21
django-extensions==4.1
django-redis==5.4.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
fonttools==4.57.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==5.3.2
MarkupSafe==3.0.2
msgpack==1.1.0
mysqlclient==2.2.7
ndg-httpsclient==0.5.1
numpy==2.2.4
oauthlib==3.3.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pusher==3.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyMySQL==1.1.1
PyNaCl==1.6.1
pyOpenSSL==25.1.0
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.1
redis==6.0.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.1
tzdata==2025.1
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.9.0
xhtml2pdf==0.2.17
zope.interface==7.2
zopfli==0.2.3.post1
